--- tests/functional/s/singledispatch/singledispatch_method_py37.rc
@@ -1,2 +1,0 @@
-[main]
-py-version=3.7
